method for editing has been setup
