+++ ConnectHub/src/main/java/com/frankaboagye/connecthub/controllers/CompanyController.java
@@ -230,49 +230,7 @@ public class CompanyController {
         return "redirect:/companyHomepage";
     }
 
-    @GetMapping("/post-a-project")
-    public String postAProject(){
-        return "postProject";
-    }
 
-    @PostMapping("/handle-post-a-project")
-    public String handleProjectPosting(
-            @ModelAttribute ProjectDAO projectDAO,
-            ModelMap modelMap,
-            HttpSession httpSession,
-            @RequestParam("documentFile") MultipartFile documentFile
-    ) {
-        // add securuty stuffs later, converstion stuffs
-
-        String stop = "here";
-
-        var date = LocalDate.parse(projectDAO.getDeadline());
-
-        // use cisco id for now
-
-        // storageServiceImplementation.store(documentFile); // commented out for the purpose of testing
-
-        // convert form dao to the object
-        Project project = Project.builder()
-                .companyId(getCisco().getId())
-                .title(projectDAO.getTitle())
-                .description(projectDAO.getDescription())
-                .skills(projectDAO.getSkills())
-                .deadline(date)
-                .location(projectDAO.getLocation())
-                .documentName(documentFile != null ? documentFile.getOriginalFilename() : "a-file-name")
-                .documentUrl("default")
-                .build();
-
-
-        companyServiceImplementation.postAProject(project);
-
-        modelMap.addAttribute("compnayProject", project);
-        modelMap.addAttribute("company", getCisco());
-        modelMap.addAttribute("SessionData", getCisco().getEmail());
-
-        return "redirect:/companyHomepage";
-    }
